rename and reorganise default configs and cellfinder main workflow script
